ScatteredOut type

ScatteringTransform.Scattered — Type

Scattered{T,N}

The abstract parent type for ScatteredOut and ScatteredFull. T gives the element type for the matrices, while N gives one plus the depth of the scattering transform (so for two layers it is three).

ScatteringTransform.ScatteredOut — Type

A simple wrapper for the results of the scattering transform. Its one field result contains a tuple of the results from each layer, in the order zero, one, two. Say we have an example s. You can access layer i by s[i], so s[0] gives the zeroth layer

ScatteredOut Utilities

ScatteringTransform.flatten — Function

flatten(scatRes) -> output

given scatRes, a scattered output or full, it produces a single vector containing the entire transform in order, i.e. the same format as output by thinSt.

ScatteringTransform.roll — Function

roll(toRoll, st::stFlux)

Given a scattering transform st and an array toRoll that is NCoeffs×extraDims, "roll" up toRoll into a ScatteredOut.

ScatteringTransform.importantCoords — Function

importantCoords(scatRes)

given a ScatteredOut scatRes, make a list that gives the largest value on each path.